+++ test/vitest/permission-store.spec.ts
@@ -1,4 +1,8 @@
-import { PermissionStore, createPermissionStore } from "fake-permissions";
+import {
+  PermissionStore,
+  createPermissionStore,
+  type HandleAccessRequest,
+} from "fake-permissions";
 import {
   afterEach,
   beforeEach,
@@ -113,6 +117,35 @@ describe("PermissionStore()", () => {
     });
   });
 
+  describe("selectByDescriptor()", () => {
+    it("selects from any iterable keyed on permission descriptors", () => {
+      const iterable = new Map([
+        [structuredClone(geolocation), "a"],
+        [structuredClone(midiSysexFalse), "b"],
+        [structuredClone(midiSysexTrue), "c"],
+      ]);
+
+      expect(
+        permissionStore.selectByDescriptor(
+          iterable,
+          structuredClone(geolocation),
+        ),
+      ).toBe("a");
+      expect(
+        permissionStore.selectByDescriptor(
+          iterable,
+          structuredClone(midiSysexFalse),
+        ),
+      ).toBe("b");
+      expect(
+        permissionStore.selectByDescriptor(
+          iterable,
+          structuredClone(midiSysexTrue),
+        ),
+      ).toBe("c");
+    });
+  });
+
   describe("getState()", () => {
     describe("when called with an unknown descriptor", () => {
       it("throws a TypeError", () => {
@@ -183,6 +216,280 @@ describe("PermissionStore()", () => {
         permissionStore.setState(geolocationWithExtra, "prompt");
 
         expect(permissionStore.getState(geolocation)).toBe("prompt");
+      });
+    });
+  });
+
+  describe("requestAccess()", () => {
+    describe("when called with an unknown descriptor", () => {
+      it("throws a TypeError", async () => {
+        const call = (async () => {
+          await permissionStore.requestAccess(notifications);
+        })();
+
+        await expect(call).rejects.toThrow(TypeError);
+        await expect(call).rejects.toThrow(
+          'No permission state for descriptor {"name":"notifications"}',
+        );
+      });
+    });
+
+    describe("when no access request handler is configured", () => {
+      describe('when access is requested for a permission in the "prompt" state', () => {
+        it("denies access and leaves the permission unchanged", async () => {
+          expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+            false,
+          );
+          expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+        });
+
+        describe("when the dialog is dismissed repeatedly", () => {
+          beforeEach(async () => {
+            await permissionStore.requestAccess(midiSysexTrue);
+            await permissionStore.requestAccess(midiSysexTrue);
+            await permissionStore.requestAccess(midiSysexTrue);
+          });
+
+          it("denies the permission automatically", () => {
+            expect(permissionStore.getState(midiSysexTrue)).toBe("denied");
+          });
+
+          it("doesn't affect other permissions", async () => {
+            expect(
+              await permissionStore.requestAccess(pushUserVisibleOnlyTrue),
+            ).toBe(false);
+            expect(permissionStore.getState(pushUserVisibleOnlyTrue)).toBe(
+              "prompt",
+            );
+          });
+
+          describe("when the permission is reset", () => {
+            beforeEach(() => {
+              permissionStore.setState(midiSysexTrue, "prompt");
+            });
+
+            it("resets the dismissal count", async () => {
+              await permissionStore.requestAccess(midiSysexTrue);
+              await permissionStore.requestAccess(midiSysexTrue);
+
+              expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+            });
+          });
+        });
+      });
+
+      describe('when access is requested for a permission in the "granted" state', () => {
+        it("allows access and leaves the permission unchanged", async () => {
+          expect(await permissionStore.requestAccess(midiSysexFalse)).toBe(
+            true,
+          );
+          expect(permissionStore.getState(midiSysexFalse)).toBe("granted");
+        });
+      });
+
+      describe('when access is requested for a permission in the "denied" state', () => {
+        it("denies access and leaves the permission unchanged", async () => {
+          expect(await permissionStore.requestAccess(geolocation)).toBe(false);
+          expect(permissionStore.getState(geolocation)).toBe("denied");
+        });
+      });
+    });
+
+    describe("when an access request handler is configured", () => {
+      let handleAccessRequest: Mock<HandleAccessRequest>;
+
+      beforeEach(() => {
+        handleAccessRequest = vi.fn(async (dialog) => {
+          dialog.allow(true);
+        });
+
+        permissionStore.setAccessRequestHandler(handleAccessRequest);
+      });
+
+      it("can't dismiss the dialog after it's been dismissed", async () => {
+        handleAccessRequest.mockImplementation(async (dialog) => {
+          dialog.allow(true);
+          dialog.deny(true);
+        });
+
+        await expect(
+          permissionStore.requestAccess(midiSysexTrue),
+        ).rejects.toThrow("Access dialog already dismissed");
+      });
+
+      describe('when access is requested for a permission in the "prompt" state', () => {
+        it("calls the callback with a dialog and the permission descriptor", async () => {
+          await permissionStore.requestAccess(midiSysexTrue);
+
+          expect(handleAccessRequest).toBeCalledWith(
+            expect.objectContaining({
+              dismiss: expect.any(Function) as () => void,
+              allow: expect.any(Function) as () => void,
+              deny: expect.any(Function) as () => void,
+            }),
+            midiSysexTrue,
+          );
+        });
+
+        describe("when the dialog is dismissed", () => {
+          beforeEach(() => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.dismiss();
+            });
+          });
+
+          it("denies access and leaves the permission unchanged", async () => {
+            expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+              false,
+            );
+            expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+          });
+        });
+
+        describe("when the dialog is dismissed repeatedly", () => {
+          beforeEach(async () => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.dismiss();
+            });
+
+            await permissionStore.requestAccess(midiSysexTrue);
+            await permissionStore.requestAccess(midiSysexTrue);
+            await permissionStore.requestAccess(midiSysexTrue);
+          });
+
+          it("denies the permission automatically", () => {
+            expect(permissionStore.getState(midiSysexTrue)).toBe("denied");
+          });
+
+          it("doesn't affect other permissions", async () => {
+            expect(
+              await permissionStore.requestAccess(pushUserVisibleOnlyTrue),
+            ).toBe(false);
+            expect(permissionStore.getState(pushUserVisibleOnlyTrue)).toBe(
+              "prompt",
+            );
+          });
+
+          describe("when the permission is reset", () => {
+            beforeEach(() => {
+              permissionStore.setState(midiSysexTrue, "prompt");
+            });
+
+            it("resets the dismissal count", async () => {
+              await permissionStore.requestAccess(midiSysexTrue);
+              await permissionStore.requestAccess(midiSysexTrue);
+
+              expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+            });
+          });
+        });
+
+        describe("when access is allowed but the decision is not remembered", () => {
+          beforeEach(() => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.allow(false);
+            });
+          });
+
+          it("allows access and leaves the permission unchanged", async () => {
+            expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+              true,
+            );
+            expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+          });
+        });
+
+        describe("when access is allowed and the decision is remembered", () => {
+          beforeEach(() => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.allow(true);
+            });
+          });
+
+          it("allows access and grants the permission", async () => {
+            expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+              true,
+            );
+            expect(permissionStore.getState(midiSysexTrue)).toBe("granted");
+          });
+        });
+
+        describe("when access is denied but the decision is not remembered", () => {
+          beforeEach(() => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.deny(false);
+            });
+          });
+
+          it("denies access and leaves the permission unchanged", async () => {
+            expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+              false,
+            );
+            expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+          });
+        });
+
+        describe("when access is denied and the decision is remembered", () => {
+          beforeEach(() => {
+            handleAccessRequest.mockImplementation(async (dialog) => {
+              dialog.deny(true);
+            });
+          });
+
+          it("denies access and denies the permission", async () => {
+            expect(await permissionStore.requestAccess(midiSysexTrue)).toBe(
+              false,
+            );
+            expect(permissionStore.getState(midiSysexTrue)).toBe("denied");
+          });
+        });
+      });
+
+      describe('when access is requested for a permission in the "granted" state', () => {
+        it("does not call the callback", async () => {
+          await permissionStore.requestAccess(midiSysexFalse);
+
+          expect(handleAccessRequest).not.toBeCalled();
+        });
+
+        it("allows access and leaves the permission unchanged", async () => {
+          expect(await permissionStore.requestAccess(midiSysexFalse)).toBe(
+            true,
+          );
+          expect(permissionStore.getState(midiSysexFalse)).toBe("granted");
+        });
+      });
+
+      describe('when access is requested for a permission in the "denied" state', () => {
+        it("does not call the callback", async () => {
+          await permissionStore.requestAccess(geolocation);
+
+          expect(handleAccessRequest).not.toBeCalled();
+        });
+
+        it("denies access and leaves the permission unchanged", async () => {
+          expect(await permissionStore.requestAccess(geolocation)).toBe(false);
+          expect(permissionStore.getState(geolocation)).toBe("denied");
+        });
+      });
+    });
+
+    describe("when a custom dismissal deny threshold is configured", () => {
+      beforeEach(() => {
+        permissionStore = createPermissionStore({
+          dismissDenyThreshold: 2,
+          initialStates: new Map([[midiSysexTrue, "prompt"]]),
+        });
+      });
+
+      it("affects how many dismissed dialogs will cause permission denial", async () => {
+        await permissionStore.requestAccess(midiSysexTrue);
+
+        expect(permissionStore.getState(midiSysexTrue)).toBe("prompt");
+
+        await permissionStore.requestAccess(midiSysexTrue);
+
+        expect(permissionStore.getState(midiSysexTrue)).toBe("denied");
       });
     });
   });